+++ include/lib/m4a/shared.h
@@ -1,0 +1,91 @@
+#ifndef GUARD_M4A_SHARED_H
+#define GUARD_M4A_SHARED_H
+
+#include "gba/gba.h"
+
+struct MP2KTrack;
+struct MixerSource {
+    u8 status;
+    u8 type;
+    u8 rightVol;
+    u8 leftVol;
+
+    union {
+        struct {
+            u8 attack;
+            u8 decay;
+            u8 sustain;
+            u8 release;
+            u8 key;
+            u8 envelopeVol;
+
+            u8 envelopeGoal;
+            u8 envelopeCtr;
+
+            u8 echoVol;
+            u8 echoLen;
+            u8 padding1;
+            u8 padding2;
+            u8 gateTime;
+            u8 untransposedKey;
+            u8 velocity;
+            u8 priority;
+            u8 rhythmPan;
+            u8 padding3;
+            u8 padding4;
+            u8 padding5;
+
+            u8 padding6;
+            u8 sustainGoal;
+            u8 nrx4;
+            u8 pan;
+
+            u8 panMask;
+            u8 cgbStatus;
+            u8 length;
+            u8 sweep;
+
+            u32 freq;
+        } cgb;
+        struct {
+            u8 attack;
+            u8 decay;
+            u8 sustain;
+            u8 release;
+            u8 key;
+            u8 envelopeVol;
+
+            u8 envelopeVolR;
+            u8 envelopeVolL;
+
+            u8 echoVol;
+            u8 echoLen;
+            u8 padding1;
+            u8 padding2;
+            u8 gateTime;
+            u8 untransposedKey;
+            u8 velocity;
+            u8 priority;
+            u8 rhythmPan;
+            u8 padding3;
+            u8 padding4;
+            u8 padding5;
+
+            u32 ct;
+            float fw;
+
+            u32 freq;
+        } sound;
+    } data;
+
+    void *wav; // The next peice of data to be loading
+    void *current; // The working pointer from wav
+
+    struct MP2KTrack *track;
+    struct MixerSource *prev;
+    struct MixerSource *next;
+    u32 padding7; // d4
+    u32 blockCount; // bdpcm block count
+};
+
+#endif // GUARD_M4A_SHARED_H
