+++ include/lib/m4a/sound_mixer.h
@@ -9,87 +9,6 @@
 
 struct MP2KPlayerState;
 
-struct MixerSource {
-    u8 status;
-    u8 type;
-    u8 rightVol;
-    u8 leftVol;
-    u8 attack;
-    u8 decay;
-    u8 sustain;
-    u8 release;
-
-    union {
-        struct {
-            u8 key;
-            u8 envelopeVol;
-            u8 envelopeCtr;
-            u8 envelopeGoal;
-
-            u8 echoVol;
-            u8 echoLen;
-            u8 padding1;
-            u8 padding2;
-            u8 gateTime;
-            u8 untransposedKey;
-            u8 velocity;
-            u8 priority;
-            u8 rhythmPan;
-            u8 padding3;
-            u8 padding4;
-            u8 padding5;
-
-            u8 padding6;
-            u8 sustainGoal;
-            u8 nrx4;
-            u8 pan;
-
-            u8 panMask;
-            u8 cgbStatus;
-            u8 length;
-            u8 sweep;
-
-            u32 freq;
-
-            u32 *newCgb3Sample; // nextSample
-            u32 *oldCgb3Sample; // currentSample
-        } cgb;
-        struct {
-            u8 key;
-            u8 envelopeVol;
-            u8 envelopeVolR;
-            u8 envelopeVolL;
-
-            u8 echoVol;
-            u8 echoLen;
-            u8 padding1;
-            u8 padding2;
-            u8 gateTime;
-            u8 untransposedKey;
-            u8 velocity;
-            u8 priority;
-            u8 rhythmPan;
-            u8 padding3;
-            u8 padding4;
-            u8 padding5;
-
-            u32 ct;
-            float fw;
-
-            u32 freq;
-
-            struct WaveData *wav;
-            s8 *current; // todo: sample
-        } sound;
-    } data;
-
-    struct MP2KTrack *track;
-    struct MixerSource *prev;
-    struct MixerSource *next;
-    u32 padding7; // d4
-    u32 blockCount; // bdpcm block count
-};
-
 enum { MAX_SAMPLE_CHANNELS = 12 };
 enum { MIXED_AUDIO_BUFFER_SIZE = 4907 };
